+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
-## Unreleased
+### [0.9.5] - [2022-05-26]
 
-* Move away from deprecated class ConsumerRecordFactory
+* Move away from deprecated class ConsumerRecordFactory (to prepare migration to Kafka Streams 3.2.0)
 
 ### [0.9.4] - [2022-05-23]